Move tab YOU to a separate single activity
We won’t use it yet though. Just moving it to a separate activity until
we finally decide whether we just drop using google location services
or we give this option to the user again.
